+++ src/VaultSharp/V1/SecretsEngines/Identity/CreateRoleRequest.cs
@@ -1,0 +1,7 @@
+﻿namespace VaultSharp.V1.SecretsEngines.Identity
+{
+    public class CreateRoleRequest : RoleInfo
+    {
+
+    }
+}