+++ src/VaultSharp/V1/SecretsEngines/Identity/ReadNamedKeyResponse.cs
@@ -1,0 +1,10 @@
+﻿using System.Text.Json.Serialization;
+
+namespace VaultSharp.V1.SecretsEngines.Identity
+{
+    public class ReadNamedKeyResponse 
+    {
+        [JsonPropertyName("data")]
+        public NamedKeyInfo Data { get; set; }
+    }
+}